Oppgrader react-pdf til versjon 10!
