+++ package.json
@@ -37,7 +37,7 @@
     "lodash.debounce": "4.0.8",
     "react": "19.2.6",
     "react-dom": "19.2.6",
-    "react-pdf": "9.2.1"
+    "react-pdf": "10.4.1"
   },
   "devDependencies": {
     "@eslint/js": "9.39.4",
