+++ src/component/pdf-viewer/pdf-status.ts
@@ -1,0 +1,6 @@
+export enum PDFStatus {
+	NOT_STARTED = 'NOT_STARTED',
+	LOADING = 'LOADING',
+	SUCCESS = 'SUCCESS',
+	ERROR = 'ERROR'
+}
